Added support for buffer sizes > Max Int

+++ sdk/storage/Azure.Storage.Blobs/api/Azure.Storage.Blobs.netstandard2.0.cs
@@ -204,7 +204,7 @@ namespace Azure.Storage.Blobs.Models
     public partial class AppendBlobOpenWriteOptions
     {
         public AppendBlobOpenWriteOptions() { }
-        public int BufferSize { get { throw null; } set { } }
+        public long BufferSize { get { throw null; } set { } }
         public Azure.Storage.Blobs.Models.AppendBlobRequestConditions Conditions { get { throw null; } set { } }
         public bool Overwrite { get { throw null; } set { } }
         public System.IProgress<long> ProgressHandler { get { throw null; } set { } }
@@ -888,7 +888,7 @@ namespace Azure.Storage.Blobs.Models
     public partial class BlockBlobOpenWriteOptions
     {
         public BlockBlobOpenWriteOptions() { }
-        public int BufferSize { get { throw null; } set { } }
+        public long BufferSize { get { throw null; } set { } }
         public Azure.Storage.Blobs.Models.BlobRequestConditions Conditions { get { throw null; } set { } }
         public bool Overwrite { get { throw null; } set { } }
         public System.IProgress<long> ProgressHandler { get { throw null; } set { } }
@@ -1060,7 +1060,7 @@ namespace Azure.Storage.Blobs.Models
     public partial class PageBlobOpenWriteOptions
     {
         public PageBlobOpenWriteOptions() { }
-        public int BufferSize { get { throw null; } set { } }
+        public long BufferSize { get { throw null; } set { } }
         public Azure.Storage.Blobs.Models.PageBlobRequestConditions Conditions { get { throw null; } set { } }
         public bool Overwrite { get { throw null; } set { } }
         public System.IProgress<long> ProgressHandler { get { throw null; } set { } }
