+++ sdk/storage/Azure.Storage.Blobs/src/Models/BlockBlobOpenWriteOptions.cs
@@ -17,10 +17,10 @@ namespace Azure.Storage.Blobs.Models
         public bool Overwrite { get; set; }
 
         /// <summary>
-        /// The size of the buffer to use.  Default is 1 MB, and the max
-        /// buffer size is 100 MB.
+        /// The size of the buffer to use.  Default is 1 MB,
+        /// max is 100 MB.
         /// </summary>
-        public int BufferSize { get; set; }
+        public long BufferSize { get; set; }
 
         /// <summary>
         /// Optional <see cref="BlobRequestConditions"/> to add
